refactor: 공통으로 사용되는 변수, 혹은 값들 -> CommonRestAssured 클래스로 분리
