+++ src/main/java/kuit/subway/domain/Station.java
@@ -16,6 +16,9 @@ public class Station {
     @Column(length = 20, nullable = false)
     private String name;
 
+
+    private Line line;
+
     public Station(String name) {
         this.name = name;
     }
